Support both local and cloud endpoints

+++ docker/run-unimrcp-demo.bash
@@ -12,6 +12,7 @@ UNIMRCP_CR="${UNIMRCP_CR:-skyman.azurecr.io}"
 UNIMRCP_REPO="${UNIMRCP_REPO:-scratch/onprem/unimrcp}"
 UNIMRCP_CLIENT_TAG="${UNIMRCP_CLIENT_TAG:-asrclient}"
 UNIMRCP_SERVER_TAG="${UNIMRCP_SERVER_TAG:-server}"
+SPEECH_SELECT="${SPEECH_SELECT:-local}"
 
 # parsing command line arguments:
 while [[ $# > 0 ]]; do
@@ -22,6 +23,7 @@ while [[ $# > 0 ]]; do
             echo "Options:"
             echo "  -r|--region <region> - speech region to use"
             echo "  -k|--apikey <key> - api key to use"
+            echo "  -c|--cloud - select cloud instead of local container"
             exit 1
             ;;
         -r|--region)
@@ -31,6 +33,9 @@ while [[ $# > 0 ]]; do
         -k|--apikey)
             SPEECH_APIKEY="$2"
             shift
+            ;;
+        -c|--cloud)
+            SPEECH_SELECT="cloud"
             ;;
         *)
             echo "Unknown option $key -- Ignoring"
@@ -46,7 +51,11 @@ SPEECH_ENDPOINT="${SPEECH_CONTAINER_NAME}:${SPEECH_CONTAINER_PORT}"
 SPEECH_DCR="${SPEECH_CR}/${SPEECH_REPO}:${SPEECH_TAG}"
 UNIMRCP_SERVER_DCR="${UNIMRCP_CR}/${UNIMRCP_REPO}:${UNIMRCP_SERVER_TAG}"
 UNIMRCP_CLIENT_DCR="${UNIMRCP_CR}/${UNIMRCP_REPO}:${UNIMRCP_CLIENT_TAG}"
-sed "s/YourSpeechContainerEndPoint/$SPEECH_ENDPOINT/g" ${SOURCE_DIR}/config.json > ${SOURCE_DIR}/config.json.rej
+sed -e "s/YourSpeechContainerEndPoint/$SPEECH_ENDPOINT/g" -e "s/YourSubscriptionKey/$SPEECH_APIKEY/g" -e "s/YourServiceRegion/$SPEECH_REGION/g"  ${SOURCE_DIR}/config.json > ${SOURCE_DIR}/config.json.rej
+
+if [[ "${SPEECH_SELECT}" == "cloud" ]]; then
+    sed -i "s/sr_use_local_container\"\:\strue/sr_use_local_container\": false/g" ${SOURCE_DIR}/config.json.rej
+fi
 
 # Cleanup previous instances of the containers
 docker kill ${SPEECH_CONTAINER_NAME} unimrcpserver asrclient || echo "No containers running, that's OK"
@@ -63,11 +72,16 @@ if [[ "${SPEECH_REGION}" == "invalid" ]]; then
 fi
 
 # Run docker images
-docker pull ${SPEECH_DCR}
+DOCKER_LINK=""
+if [[ "${SPEECH_SELECT}" != "cloud" ]]; then
+    docker pull ${SPEECH_DCR}
+    docker run --rm -d --name ${SPEECH_CONTAINER_NAME} -e DECODER_MAX_COUNT=${SPEECH_MAX_DECODERS} -p ${SPEECH_CONTAINER_PORT}:5000 "${SPEECH_DCR}" eula=accept billing=${SPEECH_BILLING} apikey=${SPEECH_APIKEY}
+    DOCKER_LINK="--link ${SPEECH_CONTAINER_NAME}:${SPEECH_CONTAINER_NAME}"
+fi
+
 docker pull ${UNIMRCP_SERVER_DCR}
 docker pull ${UNIMRCP_CLIENT_DCR}
-docker run --rm -d --name ${SPEECH_CONTAINER_NAME} -e DECODER_MAX_COUNT=${SPEECH_MAX_DECODERS} -p ${SPEECH_CONTAINER_PORT}:5000 "${SPEECH_DCR}" eula=accept billing=${SPEECH_BILLING} apikey=${SPEECH_APIKEY}
-docker run --rm -d --name unimrcpserver --link ${SPEECH_CONTAINER_NAME}:${SPEECH_CONTAINER_NAME} -v ${SOURCE_DIR}/config.json.rej:/usr/local/unimrcp/conf/config.json -p 8060:8060/udp -p 8060:8060/tcp -p 1544:1544 -p 1554:1554 "${UNIMRCP_SERVER_DCR}"
+docker run --rm -d --name unimrcpserver ${DOCKER_LINK} -v ${SOURCE_DIR}/config.json.rej:/usr/local/unimrcp/conf/config.json -p 8060:8060/udp -p 8060:8060/tcp -p 1544:1544 -p 1554:1554 "${UNIMRCP_SERVER_DCR}"
 
 echo "RUN THE FOLLOWING COMMAND> run grammar.xml whatstheweatherlike.wav"
 docker run --rm -ti --name asrclient --net=host "${UNIMRCP_CLIENT_DCR}"
